requirements.txt
docopt
deso
emoji
python-box
pyspellchecker
